moved `components` directly into `src`, not `src/app`
