+++ new-dti-website-redesign/src/app/apply/page.tsx
@@ -1,4 +1,4 @@
-import Layout from '../components/Layout';
+import Layout from '../../components/Layout';
 
 export const metadata = {
   title: 'DTI APPLY PAGE',
