+++ new-dti-website-redesign/src/app/initiatives/page.tsx
@@ -1,4 +1,4 @@
-import Layout from '../components/Layout';
+import Layout from '../../components/Layout';
 
 export const metadata = {
   title: 'DTI INITIATIVES PAGE',
