+++ new-dti-website-redesign/src/app/page.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import Layout from './components/Layout';
+import Layout from '../components/Layout';
 
 export const metadata = {
   title: 'DIT HOMEPAGE',
