+++ new-dti-website-redesign/src/app/sponsor/page.tsx
@@ -1,4 +1,4 @@
-import Layout from '../components/Layout';
+import Layout from '../../components/Layout';
 
 export const metadata = {
   title: 'DTI SPONSOR PAGE',
