+++ new-dti-website-redesign/src/app/test-components/page.tsx
@@ -2,11 +2,11 @@
 
 import Link from 'next/link';
 import { Plus } from 'lucide-react';
-import Button from '../components/Button';
-import IconButton from '../components/IconButton';
-import Input from '../components/Input';
-import LabeledInput from '../components/LabeledInput';
-import TestimonialCard from '../components/TestimonialCard';
+import Button from '../../components/Button';
+import IconButton from '../../components/IconButton';
+import Input from '../../components/Input';
+import LabeledInput from '../../components/LabeledInput';
+import TestimonialCard from '../../components/TestimonialCard';
 
 export default function TestComponents() {
   return (
