+++ new-dti-website-redesign/src/app/test-page/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import Hero from '../components/Hero';
-import Layout from '../components/Layout';
+import Hero from '../../components/Hero';
+import Layout from '../../components/Layout';
 
 export default function TestPage() {
   return (
