+++ new-dti-website-redesign/src/components/Hero.tsx
